// src/weather/weather.interface.ts
export interface Weather {
    
    T2M: Record<string, number>;   

    T2M_MAX: Record<string, number>;   

    T2M_MIN: Record<string, number>;   

    RH2M: Record<string, number>;   

    TS: Record<string, number>;   

    PRECTOTCORR: Record<string, number>;  
    
    PS: Record<string, number>;   

    WS10M: Record<string, number>;   

    WS10M_MAX: Record<string, number>;   

    WS10M_MIN: Record<string, number>;   

  }
  